limit /reset to be used in dev environment

+++ reset.go
@@ -3,6 +3,11 @@ package main
 import "net/http"
 
 func (cfg *apiConfig) handlerResetMetrics(w http.ResponseWriter, r *http.Request) {
+	if cfg.platform != "dev" {
+		w.WriteHeader(http.StatusForbidden)
+		return
+	}
+
 	err := cfg.db.DeleteAllUsers(r.Context())
 	if err != nil {
 		respondWithError(w, http.StatusInternalServerError, "couldn't delete all users", err)
